requirements.txt
Django==3.2
django-quill-editor==0.1.22
